Addressing Gna bug 21391 https://gna.org/bugs/?21391
Original problem description:
Biblatex expects the keywords to be separated by comma. With semicolon
(as currently written by KBibTeX), keyword parsing e.g. with \ifkeyword
does not work.
I propose an option to customize the separator (sticking with semicolon
as default, but providing comma as alternative). Since KBibTeX already
reads both "," and ";", a change is only necessary for export.

A patch [1] was provided by the bug reporter Jurgen Spitzmuller. This patch
was revised and refactored by Thomas Fischer resulting into this commit.

[1] https://gna.org/support/download.php?file_id=20108