Propagating commit d0ba905's changes through class
Propagating changes made in the previous commit (to address a specific
bug) to similar cases in the same class:
To get the BibTeX element behind a QTreeView item, the model index has
to be mapped through a filter/sort proxy. A function for that purpose
was already existing, but wasn't not used in all applicable cases.