Value list's menu label changes for adding/changing values
As requested in comment 4 of Gna! bug 21870 [1] ("the menu entry should
make it clear whether the value is appended or replaced"), this commit
changes the label of the context menu item of the value list depending
on if the current value list value is added to the selected elements in
the main list or if the value replaces the selected elements' values.

Patch [2] was supplied by Jürgen Spitzmüller and committed with minor
modifications by Thomas Fischer <fischer@unix-ag.uni-kl.de>.

[1] http://gna.org/bugs/?21870#comment4
[2] http://gna.org/support/download.php?file_id=20594