GNA Bug 22418: Relative paths fail to get resolved
Forward-port from branch kbibtex/0.5, commit 66e91b36373e4d7b8e1c50

As reported as GNA bug 22418 [1], opening associated files (like PDF
files linked to an entry) fail (a) if the filename given in relative
form and (b) the working directory does not match the .bib file's
directory.

The original patch (GNA patch 21615 [2]) attempted to fix the issue but
broke loading files specified with an absolute file path.

The fix as of this commit joins the .bib file's current directory (not
necessarily the CWD) and the filename is the latter is relative. In any
other case, the linked filename is used as-is.

[1] https://gna.org/bugs/?22418
[2] https://gna.org/bugs/download.php?file_id=21615