Fixing regression introduced by refactoring filter bar
Commit 0d79f0c5f8b64e broke with the assumption that the KPart's widget
is always of class FileView. As it is now a plain QWidget acting as a
container for multiple objects, various dynamic and object casts return
NULL.
This commit tries to address this issue, by using a new class for the
KPart's widget, which allows to extract the FileView object in a similar
manner as was possible earlier.