Fixing ScienceDirect search
As reported in KDE bug 339086, ScienceDirect search is no longer
working for version 0.5.1. It seems like this is due to changes
on their webpage, breaking KBibTeX's HTML code parser.
This patch adopts the parser to compensate for this issue.

An improvement over the previous implementation is that now an
intermediate step is skipped, thus less HTTP requests per
bibliography entry are necessary.