Avoid putting objects on stack that get deleted
When performign a 'save as' operation, for some export formats
(BibTeX, ...) a settings dialog may get shown where users can
specify the exact parameter of the file to be executed.
The main widgets of those dialogs were created on the stack
(plain variables, no 'new' or '*'). However, due to the parent
relation with the dialog (created the heap using 'new'), they
were destroyed automatically when the stack was cleaned up,
but also when the dialog got destroyed.

The change introduced in this patch avoids this issue by creating
the dialogs' main widgets on the heap instead ('... = new ...').