Number of selected lines was wrongfully determined
Manual backward port of commit d70f5f271c516e9b1d5 from master.
Original description reads as follows:

This commit fixes a line of code where the number of selected lines
is determined. The line of numbers is used for the duplicate finding
process to ask the use if only selected lines should be checked for
duplicates, or if all entries are to be tested.