RB 121146: extract file tests in an own kbibtexfilestest
Pino Toscano <pino@kde.org> writes:

Extract the file import/export tests from the (GUI) interactive
kbibtextest in an own non-interactive kbibtexfilestest, which does not
save files other than the temporary files for saving testing.

This makes it possible to run these tests with the usual `make test`.

The step following this patch is to remove the file tests from
kbibtextest, would that be ok?