Message next to Import button in Search Results
As suggested in bug report 344497, an inline message
(an icon and a label) is shown next to the Import
button in the Search Results docklet if this button
is disabled.
There are two reasons, why the button may be disabled:
1. No results to import.
2. No bibliography open to import to.

Patch 91251 provided by Arseniy Lartsev (original bug
reporter) did only cover case 2, so it had to be
rewritten and expanded.

BUG: 344497