Migrating from QCA2 to QCA-Qt5
The previously used CMake module FindQCA2 is owned
by KDE4, which should be avoided in this Frameworks
port.
The proper replacement seems to be QCA-Qt5.
Unfortunately, this package doesn't seem to
set an include directory where to find QtCrypto.
Therefore, new code had to be included to make
an educated guess where to find the proper
include directory.