Supporting Zotero's rate limiting
Adding basic support for Zotero's rate limiting
which comes into effect if
- Zotero's servers are overloaded
- the client performs too many requests per time
In both cases, Zotero tells the client how many
seconds the client shall wait until the next
request.

Using this patch, new requests to Zotero will be
delayed if a previous answer included a request
to backoff.

Implemented using QTimer::singleShot and functors/
lambdas (?), which is a first time in KBibTeX.