Deferring loading .rc file
For KBibTeX/KDE4 versions, a recent refactoring of the
constructor and related functions (commit ac3cc2bc74924)
caused the .rc file that describes the GUI (esp. the
menu structure) was loaded too early and thus did not
come into proper use.

Although it does not seem to be a problem for the KF5
build, it seems wise to apply a similar commit here
to avoid potential future problems.

Effectively, this is a manual forward-port of commit
2a46741d2a2d4c754.

Bug 354160 was reported by Juergen Spitzmueller.