Use only QTextDocument/QTextEdit for reference preview
The reference preview uses a HTML rendering engine to show
references in a pretty fashion. So far, either a WebKit,
a WebEngine, a QLabel, or a QTextDocument/QTextEdit-based
solution (depending on version, platform, and availability)
has been used.

To simplify the code and in order to make rendered HTML
output to use the same font size as the regular user
interface, QTextDocument/QTextEdit will be used
exclusively to preview references.

BUG: 354785