Colored entries get black or white foreground
An existing feature allows to assign a (background) color
to an entry. This background color may have a poor contrast
to the text color used in the list view.

This patch looks at the background color's grey value and
based on this enforces either black or white as text color
for background-colored entries.

Manual backport of commit 935dd6631e301da31a9d67.