Splitting Clipboard class, creating FileOperation
Except for functions that operate directly on the system's
clipboard or are part of drag'n'drop oerations, the Clipboard
class's code has been refactored into the new class FileOperation.

This new class performs operations such as retrieving textual
representations of selected elements or to create new entries
based on BibTeX input.

Signed-off-by: Thomas Fischer <fischer@unix-ag.uni-kl.de>