Removing file with license CC BY-NC
As discussed in bug 358446, having a file licensed under
a Creative Commons Attribution-Non-Commercial is problematic.
The dicussion led to the conclusion that having this file
in the repository is no license violation, as this text file
was never linked to GPL-code and was not installed by default,
i.e. never shipped with distributions' packages.

Still, to avoid future problems, this file is/was removed by
a simple 'git rm'. Rewriting the Git repository to purge this
file from it has been deemed as not necessary.