Default id automatically applied to new elements
This commit tries to address KDE bug 353890. The application of
the default entry id as configured by the user may not have been
automatically applied to new elements.
This commit tries to address this issue by rewriting the signals
used to detect changes in the editing dialog and user changes
to the entry id lineedit.