Correctly creating backup on symlink'ed files
This commit adds code that is necessary to handle the following
situation: User opens, modifies, and saves file /tmp/b/bbb.bib which is
actually a symlink to file /tmp/a/aaa.bib.
The new code therefore checks if /tmp/b/bbb.bib is to be copied/moved to
/tmp/b/bbb.bib~ and /tmp/b/bbb.bib is a local file and /tmp/b/bbb.bib is
a symbolic link to another file. Then /tmp/b/bbb.bib is resolved to the
real file /tmp/a/aaa.bib which is then copied into plain file
/tmp/b/bbb.bib~.
The save function (not modified by this commit) will then see that
/tmp/b/bbb.bib is a symbolic link, resolve this symlink to
/tmp/a/aaa.bib, and then write all changes to /tmp/a/aaa.bib keeping
/tmp/b/bbb.bib a link to.