Monitoring file status changes
Monitoring file status changes to get notified of 'Save As' operations
that allow users to change file properties (there is a dialog based on
the same widget as this docklet). Whenever this signal is detected,
this docklet needs to reload the properties from the current file.

This commit is a manual forward-port of commit 4736544ffafb5ca194943f
from branch 'kbibtex/0.7'.