Reordering call to setModel(..)
Calling 'setModel(..)' early in the patched function caused
some ripple effect that eventually tried to access uninitialized
data.
The patch fixes this problem by moving the call to 'setModel(..)'
further down into the function. Additionally, a warning is shown
if the concerned private variable was not set to a non-NULL value.

Suspicion: In Qt 5.7, the semantics of 'QAbstractItemView::setModel'
was changed causing updates in the model which eventually required
an access to a variable not yet initialized.