Speeding up selection of elements
The old code manually iterated over all columns of rows
to be selected, which was expensive in the case of many
columns (e.g. for BibLaTeX) and large files (many rows).

This patch uses a Qt-internal flag when selecting elements
to select the full row with a single call.

Furthermore, the unused function
'setSelectedElement(QList<QSharedPointer<Element> > &)'
has been completely removed as it was unused.

Signed-off-by: Thomas Fischer <fischer@unix-ag.uni-kl.de>

BUG: 371515