Speeding up selection of elements
The old code manually iterated over all columns of rows
to be selected, which was expensive in the case of many
columns (e.g. for BibLaTeX) and large files (many rows).

This patch uses a Qt-internal flag when selecting elements
to select the full row with a single call.

Furthermore, the unused function
'setSelectedElement(QList<QSharedPointer<Element> > &)'
has been completely removed as it was unused.

This commit fixes bug 371515 and is a manual forward-port
of commit 98e55392df85285a from branch kbibtex/0.6.

Signed-off-by: Thomas Fischer <fischer@unix-ag.uni-kl.de>