Zotero::API instance managed through QSharedPointer
In order to address issues where stale instances of Zotero::API are
used, the single instance of this class, which is used in a number
of other classes, is encapsulated in a QSharedPointer to keep track
of which other objects are still using this instance.

Manual forward-port of commit 0ac3b8a4cab0875114ce83379 from branch
'kbibtex/0.6'.