Invalid configuration causes crash in BasicFileView::Private
An invalid entry in KBibTeX's configuration file may cause a crash
during the initialization of BasicFileView::Private. The 'crash'
was caused by a failed Q_ASSERT_X.
Now the values get checked for consistency (at least the sum of
column widths). A width of zero, which is invalid, will now cause
a reset to supposedly correct default values.

BUG: 373768