Showing search results docklet on data returned from Zotero
Fixing a minor issue as reported in comment 6 of
KDE bug report 377401:

"However, the search results panel doesn't open when I
double click on the zotero library. It only shows
results if I did an online search before and when the
search panel is opened. If I close the search panel
it doesn't appear when double clicking the library."