If QtOAuth is not available, support manually entering credentials
In case the QtOAuth library was not linked into KBibTeX,
still allow the user to enter a user id and an API key
as manually retrieved from the Zotero homepage.

The newly added code will do some basic checking on the
entered credentials (e.g. format of the API key) and
make use of those credentials if they look ok. Otherwise
the user will be notified that the credentials are most
likely wrong or incomplete.