Avoid duplicate hits when extracting URLs from entries
When extracting URLs from an entry, KBibTeX first looks for
text patterns that look like DOIs, which are then prefixed
with 'http://dx.doi.org' (automatic online resolver).
To avoid multiple hits, the old code removed the extracted
DOI pattern from the input text and continued to look for
URLs in the remaining text.

However, in scenarios where the input text was like
  'http://doi.example.org/10.1051/a946-000'
the remainder after removal of the DOI would have been
  'http://doi.example.org/'
which then in a later step would have been as well
recognized as a separate URL.

The new code introduced with this commit checks if the
text directly preceeding a found DOI looks like an URL.
If that is the case, the URL-like part will be removed
from the input text before the search for more URLs
continues. Extracted DOIs will still be prefixed with
'dx.doi.org'.