Remove 'static' from QObject-derived objects
Objects of classes that are derived from QObject are supposed to be in a
clear parent-children relation: at object creation, a QObject-derived
parent is passed to the QObject constructor.
QObject's destructor will take care that all child objects will be
destroyed as well.
This conflicts with QObject-derived objects that are declared 'static',
as this may imply a different life span than the object's parent.
This may lead to double-destroy situations crashing KBibTeX.

This commit will avoid having the 'static' keyword at variable
declarations of QObject-derived objects.

BUG: 378497
FIXED-IN: 0.8