For BibTeX and PDF files, use KBibTeX's part
Ignore KDE's global settings for which KPart to use for BibTeX or
PDF files; always use KBibTeX's own KPart.

Opening PDF files will work if PDF file was saved using KBibTeX,
as it can embedd a BibTeX file into a PDF file when saving and
transparently extract it when opening this PDF file.