Handling of empty or duplicate entry ids or macro keys
Improving the handling of empty entry ids or macro keys,
respectively, or entry ids or macro keys, respectively, that
are duplicates of existing entries' ids or macros' keys,
respectively.

For empty ids/keys, a warning is shown and the original
(hopefully valid) id/key is restored.
For duplicate ids/keys, the user is asked to choose between
restoring the original id/key or keeping the new-but-duplicate
id/key.

BUG: 384741