Refactoring automated file test
Refactoring automated, QtTest-based file test to correctly work again.

Instead of hard-coding expected results, the test program can be
compiled with the definition of WRITE_RAWDATAFILE enabled to write out a
header file that contains definitions of QByteArrays with expected
results.
During normal operation, i.e. without WRITE_RAWDATAFILE, this header
file will be included and use as a baseline to see if loading and saving
of selected testset files is working correctly.