First check for qtoauth-qt5, only then for qtoauth
During the cmake phase, KBibTeX will look for a Qt-based OAuth library.
Previously, it used PkgConfig to identify a library called 'qtoauth',
assuming this would be a Qt5-based library as it is the case for
Linux distributions such as Gentoo Linux or Arch Linux.

However, other distributions such as Fedora maintain both a Qt4-based
and a Qt5-based variant of this library under the names 'qtoauth' and
'qtoauth-qt5', respectively. Thus, simply checking for a library called
'qtoauth' would result in linking to a Qt4-based library and in issues
like reported in bug 387638.

This commit introduces an advanced check, where first a test for
'qtoauth-qt5' is done. If this probe fails (like on Gentoo), a test
for 'qtoauth' is done.
Eventually, the correct Qt5-based QtOAuth library will be identified.

BUG: 387638