Replacing deprecated QxtSpanSlider with home-made widget
Removing deprecated LibQxt extension library for Qt. As it turns out,
the previously QxtSpanSlides does no longer properly react on mouse
events.

This required the introduction of a new widget, RangeWidget, which
mimics the same functionality by providing two QComboBoxes to allow
the user to select minimum and maximum values.

BUG: 389306