Adding validation code throughout widgets
Various widget and helper classes used to edit fields of
elements, preambles and comments got a 'validate(..)' function
that evaluates the widgets' current textual content for
validity as value for a BibTeX field.

The valiation functions propagate back a pointer to at least
one widget that fails a validation test if such a widget
exists (nullptr is returned otherwise). This allows the UI to
notify the user and to put focus on the widget that has
issues.