Resolving symbolic links to LyX pipe
Newer Kile versions (at least 2.9.91) create the LyX pipe files in /tmp
(on Linux systems, at least) and then place symbolic links to those pipes
in the user's home directory both as ~/.lyxpipe.in and ~/.lyx/lyxpipe.in
(for the 'in' pipe).

Before this patch, KBibTeX would detect those symbolic links and discard
them as not being named pipes instead of resolving them to get the real
named pipe.
As introduced in this patch, symbolic links are resolved using
QFileInfo::canonicalFilePath().

As another change introduced in this patch, the previous code
duplication when checking both in ~/ and ~/.lyx for a LyX pipe has been
addressed.

BUG: 393224
FIXED-IN: 0.8