Rewriting QRegExp code to non-regexp variant; adding tests
In an effort to get rid of the deprecated QRegExp class and in
order to improve the general code quality, usage of QRegExp in
OnlineSearchAbstract's HTML form parsing function has been
rewritten to not use any regular expressions at all.
The new code is test-covered using an automated test using QTest.