Migrating from QList to QSet for functions in FileInfo
For collections of URLs to be extracted from text or entries, the data
structure for URL collections has been migrated from QList to QSet.
This is motivated by the fact that no order is necessary for the
collection and QList is said to be rather inefficient (even if order is
required).