DropBox-synchronized folders no longer cause 'reload' messages
It is strongly suspected that when saving a file in a folder
synchronized via DropBox, the DropBox client modifies(*) the file
right after it was saved in KBibTeX, causing KBibTeX to ask the
user if the file was to be reloaded or on-disk changes to be
ignored.
By delaying resuming watching a local file after it has been saved to
disk by 500ms, this problem is no longer apparent.

(*) The file's content does not seem to be changed, but some filesystem
metadata may get touched/modified, triggering the file watching service.

This commit was cherry-picked from branch 'kbibtex/0.8', commit
f959986e4c582f24c4.

CCBUG: 396343