Replacing char and char* with QChar and QString
Class EncoderLaTeX heavily relied on raw char and char* to
handle single letters and strings. This caused trouble when
chars where to be compared to non-ASCII letters in QChar,
for example.
To make this class a better Qt citizen, usage of raw char
and char* has been refactored to a great extend.