Initial support for Science Direct API v2
Towards the end of 2018, Elsevier is migrating the API for ScienceDirect
from v1 to v2. Major changes include the (exclusive) use of JSON (a
first for KBibTeX's online search engines) and that queries are
preferrably issued via PUT requests rather than GET.

As the consequences of this API migration, the former XSL transformation
is no longer applicable and the class OnlineSearchScienceDirect
underwent a major rewrite.