NotificationHub used to publish bibliography system changes
Whenever the bibliography system gets changed, the NotificationHub is
used to publish information about this change. For that, a new event id
has been introduced.

First user is the FileModel which will emit a headerDataChanged signal
if it learns that different columns are to be shown after a bibliography
system change.