Optionally protect casing of title etc. in RIS importer
When importing bibliographic data from RIS files, now the caller can set
if the casing of title, book title, journal, etc. shall be protected,
i.e. if such fields shall have one of the two semantics:
  title = {{Some Title}}
or
  title = {Some Title}
In the latter case, the capital 'T' may get converted to a lower case
't' when BibTeX is invoked (KBibTeX won't touch it).