Refactoring KBibTeXPart::KBibTeXPartPrivate::getSaveFilename
To support more features, instead of accepting a single 'bool' as argument,
now flags based on an enum with fix power-of-two growing values are used.

For unclear reasons, Q_DECLARE_OPERATORS_FOR_FLAGS cannot be used. Compilation
breaks, requiring further investigation. An old-fashion but not type-safe
solution is used as a stop-gap measure.