Adding script generating CamelCase header include files
This becomes necessary as the ongoing refactoring and modernization of
CMake files and header includes migrated to CamelCase include
statements, for example:
  #include <Element>
instead of
  #include <element.h>

Whereas CMake-based build systems can make use of modules taking care of
generating CamelCase headers, qmake-based build systems do not (yet)
have this functionality.

Using this script is only a stopgap solution which is to be replaced by
a cleaner solution, for example by migrating the SailfishOS-based
'bibsearch' to make use of CMake as the build system.