To translit Unicode to ASCII, have Iconv as alternative
The default library in KBibTeX to handle transliterations, i.e.
rewriting Unicode symbols into sequences of plain ASCII characters,
is International Components for Unicode (ICU). Whereas often
already installed on most Linux systems due to other software
depending on it, the library is less likely available on Windows
or Mac machines (or harder to link against).

As a fall-back solution, this commit adds a similar implementation
based on the Iconv library. This code paths should be used if
no ICU is available, but a modern, preferrably GNU-based, Iconv
library is available.
There are rumors that the Iconv library transliterates differently
depending on the current locale. For example 'a with two dots' may
be transliterated into 'ae' in Germany but to just 'a' in other
countries.

Further testing is required to check if the Iconv code path
works on Windows or Mac systems.