To translit Unicode to ASCII, use tables from Text::Unidecode
As an alternative to using ICU or Iconv to transliterate Unicode
symbols into sequences of ASCII characters, make use of tables
shipped with the Perl module 'Text::Unidecode'.

This commit includes a Perl script on its own that parses those
tables and prints C code for inclusion into 'src/io/encoder.cpp'.
See the comments in this Perl code for further (technical) details.