When switching btwn BibTeX and BibLaTeX, reset model
Switching between BibTeX and BibLaTeX mode means that the available
fields change (BibLaTeX supports far more than BibTeX). This also
implies that the columns available/shown in the file list views change
upon switching between both modes.

To perform such a switch properly, the model needs to be resetted in
order for the view to retrieve for all cells the correct values.