Fail if view's column count does not match bibliography mode
The file views' headers have labels depending on the current
bibliography mode (either BibTeX or BibLaTeX): 'Title', 'Author or
Editor', 'Pages', ...
Thus, to work as expected, it is required that the file views have as
many columns as the current bibliography mode knows fields. Any
inconsistency has to be considered an error.