Replacing custom urlIsLocal() with QUrl::isLocalFile()
Some old comments say that QUrl::isLocalFile() is insufficient to differ
between local and remote URLs. Unfortunately, no detailed information is
available.
Even if such an insufficency existed, it is reasonable to assume that this
problem/bug has been fixed in Qt since then. As such, it should be safe to
remove UrlListEdit::urlIsLocal() and AssociatedFiles::urlIsLocal() and
replace them with QUrl::isLocalFile().