Entry editing tabs get identifiers
Configuration files describing the tabs used in the entry/element editing
widgets get a new field in order to identifier each tab by a string.
The identifier is most often a variant of the user interface string for
each tab. For example, the identifier for a tab with label 'Title' gets
identifier 'title'. Identifiers are all lower-case and use dashes instead
of spaces used in the user interface strings.