Showing info messages in entry editing tabs
The previous commit introduced information messages to the
EntryTabLayout data structure. This commit now makes use of
KMessageWidget to show one or several info messages per tab.
If no message is given, no KMessageWidget will be shown.

Clickable links in those info messages are caught, but the
actual handling is still a to-do for future commits.