Allow manual column resizing and moving
As reported in bug 413367, the ability to move and resize columns in the
main list view got removed in commit 841be480cc2230bfe7ba as the code
was considered error-prone.

This commit re-introduces the ability to move/resize columns. This
feature is disabled by default and must be first activated via the
context menu on the list view's header.
It is also possible to reset any moved or resized columns to the default
setting via another menu item in the same context menu.

BUG: 403750
BUG: 413367
FIXED-IN: 0.11