Fixing issues with translations
Fixing a crash that apparently is connected to a failed translation
of a string as reported in bug 414375.

Burkhard Lück <lueck@hube-lueck.de> provided a patch to CMakeLists.txt
which should fix this problem.

In addition to Burkhard's patch, a GUI issue was fixed so that empty
strings are not attempted to be translated leading to placeholder
messages shown instead of just nothing.

This commit is a backport from master's commit 84c5042c79d14667c5d9cd.

CCBUG: 414375
CCBUG: 414385