Fixing issues with translations
Fixing a crash that apparently is connected to a failed translation
of a string as reported in bug 414375.

Burkhard Lück <lueck@hube-lueck.de> provided a patch to CMakeLists.txt
which should fix this problem.

This commit is a manual backport from master's commit 84c5042c79d1466
regarding file CMakeLists.txt, excluding the GUI fix.

CCBUG: 414375
CCBUG: 414385
FIXED-IN: 0.9.2