Fixing issues with translations
Fixing a crash that apparently is connected to a failed translation
of a string as reported in bug 414375.

Burkhard Lück <lueck@hube-lueck.de> provided a patch to CMakeLists.txt
which should fix this problem.

In addition to Burkhard's patch, a GUI issue was fixed so that empty
strings are not attempted to be translated leading to placeholder
messages shown instead of just nothing.

BUG: 414375