Refactoring favicon retrieval
Instead of manually storing and from time to time updating the
URL of an online search's favicon, this new code locates and
retrieves the favicon only based on the online search's homepage
URL.

Central change is the introduction of the class FavIconLocator
that, only given a webpage URL, tries to determine/guess the
corresponding favicon URL. Once an URL is determine, the icon
will be downloaded and provided to the caller as a QIcon object.
All this webpage and favicon retrieval is asynchronous and
controlled via signals and slots.