Associating (new) URL with bibliography when saving
When saving a newly created bibliography to a file, set the file's
URL as a property to the bibliography object.
The URL stored in this property allows, among others, to recognize
URLs associated with individual entries, such as PDF files, as
relative to the BibTeX file. Without an URL, only absolute URLs
would be possible.

This commit is a manual forward-port of commit bde43759f26c2903bc6
from branch 'kbibtex/0.9'.

CCBUG: 409281