Associating (new) URL with bibliography when saving
When saving a newly created bibliography to a file, set the file's
URL as a property to the bibliography object.
The URL stored in this property allows, among others, to recognize
URLs associated with individual entries, such as PDF files, as
relative to the BibTeX file. Without an URL, only absolute URLs
would be possible.

BUG: 409281
FIXED-IN: 0.9.2