File URLs passed by command line preceed over LRU
At program start, bibliography files may be opened either because their
filenames/URLs were passed as command line arguments or because they
were open in the previous session.
Until now, the actual file being shown to the user was one of the
previously opened files as the list of recently opened files was read
with a small delay (got referred for unrelated reasons).

However, if a user invokes KBibTeX with a file name or URLs as argument,
it is reasonable to assume that this file should be shown directly
instead of any previously opened files (which still may get opened in
the background).

This commit also fixes the minor usability issue that if only previously
opened files are to be restored (and no command line argument was
given), the most recently used/opened bibliography file should be
preferrably shown instead of an arbitary one.

Forward-port of commit f56c52e9baad7f6e5 from branch 'kbibtex/0.10'
which in its turn is a forward-port of commit ce455dfd2ebb01b5685fb
from branch 'kbibtex/0.9'.

CCBUG: 417164