Correction on prev commit on editing comments
Correcting a previous commit that was meant to allow editing comments.
The previous commit (5aebdec1b671d9a36) fixed an issue in the editing
widget to accept changes to comments to be edited, but the changes were
never applied to the internal representation of the bibliography file.

Commit ce1874b1cbb935fa of branch 'kbibtex/0.9' is a combination of this
commit here and the earlier commit (5aebdec1b).