Confirm discarding changes when closing editing dialog
Ask the user whether changes which did not get applied shall be
discarded when the element editing dialog is closed (either by
closing the window through the window manager or by pressing
the 'Cancel' button).

This commit is a forward-port of commit 6fe7fa3a33755d9 from
branch 'kbibtex/0.9'.

CCBUG: 417210