Fixing and revamping ACM's Digital Library
Fixing and revamping ACM's Digital Library (a.k.a. ACM Portal) online search
which became necessary due to a recent redesign of the homepage. Less HTTP
requests are necessary now as the bibliography data for multiple items can be
fetched in one request. Data returned by ACM is in JSON format.