When resetting FieldLineEdit, use preferred type flag
This commit addresses the issue reported in bug 417210,
comment 4, issue (d).

Among others, MenuLineEdit and, by extension, FieldLineEdit
got proper clear() functions and now a emit a modified()
signal which is different to the textChanged(QString) signal
as the modified() signal includes changes to the type as well.

CCBUG: 417210
FIXED-IN: 0.9.2