When resetting FieldLineEdit, use preferred type flag
This commit addresses the issue reported in bug 417210,
comment 4, issue (d).

Among others, MenuLineEdit and, by extension, FieldLineEdit
got proper clear() functions and now a emit a modified()
signal which is different to the textChanged(QString) signal
as the modified() signal includes changes to the type as well.

This commit is a forward-port of commit d3f7e42afdb994 from
branch 'kbibtex/0.10' which in its turn is a forward-port of
commit 22ca454abbd7f3a from branch 'kbibtex/0.9'.

CCBUG: 417210