When suggesting entry ids, do not count 'small words'
When generating entry ids based on title or journal title, a range of
words, such as from first to fourth word can be specified.
Before this commit, 'small words', despite being removed in the id
generation process, were counted when determining the first, second, ...
word.
If the suggestion template stated 'first word only' and the first word
was a 'small word' such as 'the', an empty suggestion may have been
generated.

BUG: 421612