Neither open nor close QIODevice during save(..)
Changing from previous semantics, the QIODevice passed to save(..) functions
of FileExporter classes will only be checked if writeable, i.e. already open
in write mode, but no attempt will be done to open it inside the save(..)
functions.
Likewise, the device will no longer be closed at the end of the save(..)
functions. This allows, among others, to determine how many bytes have been
written to the device.

In summary, it is up to the caller of the save(..) functions to provide an
open device to the functions and to close it once the save(..) functions have
returned.