Aligning drop event handling of FieldListEdit and FieldLineEdit
As detected by Coverity Scan CID 359878, there was a resource leak in FieldListEdit.
An earlier commit (64cb83fa70e927fe332f5) introduced QScopedPointer to
FieldLineEdit::dropEvent(..) but the corresponding change was missing in
FieldListEdit::dropEvent(..).

This commit updates FieldListEdit::dropEvent(..) to make use of a QScopedPointer to
release a File object upon exit.

This commit also addresses the issue that QMimeData::text() does return an empty QString
even if the data's mime type is 'text/plain' and the data is not empty. Instead, the
more complex formulation of 'QString::fromUtf8(data(QStringLiteral(text/plain)))' is
used.