Splitting 'LaTeX encoding' from real character encodings
The recent refactoring of FileExporterBibTeX introduced the regression
that in URLs containing the '#' character (such as used for HTML
anchors) got 'LaTeX-encoded' into '\#' in some situations, leading to
invalid URLs.

The decision to either protect the hash symbol with a backlash or
keeping it verbatim depends on the data type of a field's value, e.g.
the verbatim code of an URL. Thus, the proper case handling must be made
when such field values are written into BibTeX code.

The choice of character encoding, such as UTF-8 or UTF-16, has to be
made on the highest level as it applies to all characters in a BibTeX
file, such as the keywords starting an entry like '@article'.

The 'split' thus happens as typical user choices for 'encoding'
including 'LaTeX', 'UTF-8', or 'ASCII' have different effects at
different places when writing/exporting a bibliography.