Addressing \textmugreek vs \textmu and Greek mu (U+03BC) vs micro (U+00B5)
The package 'textcomp' defined a '\textmu' command for use as the
'micro' symbol, but not defined any more 'text Greek' symbols. This gap
was filled by the 'textgreek' package, which defines symbols like
'\textalpha' etc, but made an exception for mu and named it
'\textmugreek' in order to avoid a conflict with 'textcomp'.
Note that both commands are available in text mode, in contrast to \mu
which is a math mode command and gives you an italic form of the Greek
mu (both 'mu' symbols from 'textcomp' and 'textgreek' are upright).

Unicode defines two almost identical symbols:
- The 'micro' symbol U+00B5 used in scientific contexts as a SI prefix
  for 10^{-6}   and
- The Greek 'mu' symbol U+03BC as part of the Greek alphabet.

As there exist visually identical symbols for 'mu' and different
commands, depending on mode and used packages, there is no universal,
clear mapping between Unicode symbols and LaTeX commands.
KBibTeX's LaTeX encoder does some guesswork what may be most suitable,
but there will always be cases where it breaks.

This commit is a backport of commit 595d7eb0395778553 from branch
'master'.