Add zbMATH Open OAI-PMH search
Adding support for zBMATH Open OAI-PMH search. The current
implementation does not make use of resumption tokens, thus search
results may be limited.
While filtering for author and publication year is done on the server
side, filtering for title or free text is done locally in the results
returned from the server. This means that if the user searches only for
title or text, essentially an empty search will be sent to zbMATH. The
result may contain arbitrary items and the chances that the text the
users searches for is included is low.
The search engine will show a popup notification for a few seconds
notifying the user about this limitation.

Note that this online search makes use of the Open API of zbMATH which,
as per statement by zbMATH, provides a more limited set of results
compared to the corresponding public web-based search:

  "[T]he API only provides a subset of the data in the zbMATH Open Web
   interface"
  https://oai.zbmath.org/

BUG: 444779